requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu116
torch==1.13.1+cu116
--extra-index-url https://download.pytorch.org/whl/cu116
torchvision==0.14.1+cu116
timm==0.4.12
tensorboardX==2.4
torchprofile==0.0.4
lmdb==1.2.1
pyarrow==5.0.0
einops==0.4.1

-f https://download.openmmlab.com/mmcv/dist/cu116/torch1.13/index.html
mmcv-full==1.7.0
mmdet==2.28.1

matplotlib
scipy
